Update manualMocks.md for index.js files
This might be a no-brainer, but I got hung up on this for a while. Between this documentation and the path used to `require` modules exported by an index.js I have tried to put the `__mocks__` folder adjacent to the folder named after the mock. I wanted to submit this as a proposal for documenting this potentially confusing part.